requirements.txt
fastapi
threads-net
uvicorn